Generalize ThreeColDocx into Writer, inherit from it
